# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeView"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeView"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/me/avatar/presign": {
      "post": {
        "operationId": "UsersController_presignAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignAvatarDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignAvatarResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/me/avatar": {
      "patch": {
        "operationId": "UsersController_updateAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmAvatarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeView"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{username}": {
      "get": {
        "operationId": "UsersController_getByUsername",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicView"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{username}/follow": {
      "post": {
        "operationId": "SocialController_follow",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowStateDto"
                }
              }
            }
          }
        },
        "tags": [
          "social"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "SocialController_unfollow",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowStateDto"
                }
              }
            }
          }
        },
        "tags": [
          "social"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "SocialController_setFavorite",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFollowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowStateDto"
                }
              }
            }
          }
        },
        "tags": [
          "social"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/me/following": {
      "get": {
        "operationId": "SocialController_following",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor opaco devuelto por la página anterior.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FollowingItemDto"
                          }
                        },
                        "nextCursor": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "social"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/me/followers": {
      "get": {
        "operationId": "SocialController_followers",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor opaco devuelto por la página anterior.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FollowerItemDto"
                          }
                        },
                        "nextCursor": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "social"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Micelio API",
    "description": "Auth, folders and file storage API for Micelio",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "FeedSettingsView": {
        "type": "object",
        "properties": {
          "layout": {
            "type": "string",
            "enum": [
              "GRID",
              "MASONRY"
            ]
          },
          "columns": {
            "type": "number"
          },
          "gap": {
            "type": "number"
          }
        },
        "required": [
          "layout",
          "columns",
          "gap"
        ]
      },
      "MeView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "followersCount": {
            "type": "number"
          },
          "followingCount": {
            "type": "number"
          },
          "viewerFollows": {
            "type": "boolean"
          },
          "followsViewer": {
            "type": "boolean"
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "feedSettings": {
            "$ref": "#/components/schemas/FeedSettingsView"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "TEACHER",
              "ADMIN",
              "SUPPORT"
            ]
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "avatarUrl",
          "isPublic",
          "followersCount",
          "followingCount",
          "viewerFollows",
          "followsViewer",
          "email",
          "role"
        ]
      },
      "FeedSettingsDto": {
        "type": "object",
        "properties": {
          "layout": {
            "type": "string",
            "enum": [
              "GRID",
              "MASONRY"
            ],
            "example": "MASONRY"
          },
          "columns": {
            "type": "number",
            "example": 3,
            "minimum": 1,
            "maximum": 6
          },
          "gap": {
            "type": "number",
            "example": 2,
            "minimum": 0,
            "maximum": 5
          }
        }
      },
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ada Lovelace"
          },
          "bio": {
            "type": "string",
            "example": "Ilustradora y animadora."
          },
          "isPublic": {
            "type": "boolean",
            "example": true
          },
          "feedSettings": {
            "$ref": "#/components/schemas/FeedSettingsDto"
          }
        }
      },
      "PresignAvatarDto": {
        "type": "object",
        "properties": {
          "mimeType": {
            "type": "string",
            "example": "image/png"
          },
          "size": {
            "type": "number",
            "example": 245678
          }
        },
        "required": [
          "mimeType",
          "size"
        ]
      },
      "PresignAvatarResponseDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "uploadUrl": {
            "type": "string",
            "description": "URL firmada para un PUT directo del binario a S3."
          },
          "expiresIn": {
            "type": "number",
            "description": "Segundos de validez de uploadUrl."
          }
        },
        "required": [
          "key",
          "uploadUrl",
          "expiresIn"
        ]
      },
      "ConfirmAvatarDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "avatars/uuid/uuid.png"
          }
        },
        "required": [
          "key"
        ]
      },
      "UserPublicView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "followersCount": {
            "type": "number"
          },
          "followingCount": {
            "type": "number"
          },
          "viewerFollows": {
            "type": "boolean"
          },
          "followsViewer": {
            "type": "boolean"
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "feedSettings": {
            "$ref": "#/components/schemas/FeedSettingsView"
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "avatarUrl",
          "isPublic",
          "followersCount",
          "followingCount",
          "viewerFollows",
          "followsViewer"
        ]
      },
      "FollowStateDto": {
        "type": "object",
        "properties": {
          "following": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          }
        },
        "required": [
          "following"
        ]
      },
      "UpdateFollowDto": {
        "type": "object",
        "properties": {
          "isFavorite": {
            "type": "boolean",
            "example": true,
            "description": "Favorito: prioridad de 12 h en el home."
          }
        },
        "required": [
          "isFavorite"
        ]
      },
      "FollowingItemDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserPublicView"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "since": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "user",
          "isFavorite",
          "since"
        ]
      },
      "FollowerItemDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserPublicView"
          },
          "since": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "user",
          "since"
        ]
      },
...
    }
  }
}